fix: 테스트 코드 위치를 수정합니다
